requirements.txt
langchain==0.0.304
pypdf==3.16.2
PyPDF2==3.0.1
python_docx==0.8.11
reportlab==4.0.5
streamlit==1.27.0
tiktoken==0.5.1
faiss-cpu==1.7.4
fonttools==4.42.1
huggingface-hub==0.17.3
langsmith==0.0.41
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
openai==0.28.1
prompt-toolkit==3.0.39
pydantic==2.4.2
pydantic_core==2.10.1
python-json-logger==2.0.7
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
transformers==4.33.3
urllib3==2.0.5
youtube-transcript-api==0.6.1